refactor: create new component category and organize
